Add script to link GEFS files from production
New script that scans backwards from the current time and creates
links in input for any cycles that are available in the production
directory but don't yet have links, so they can be processed.

The script also removed input directories once the links go stale as
runs age off of the production system.