Move in/out directories to ptmp
The input and output directories are moved from the tigge directory
to the ptmp area. This required moving the TIGGE_INPUT and
TIGGE_OUTPUT envvars up to the rocoto level to be set for all jobs
simultaneously, and the removal of some hard-coded paths.

The C code was also changed to use the enivronment variable to
eliminate the need for symlinks in the tigge directory when the
directories were located elsewhere.

Also fixed some bugs in the transfer script.