Streamline input filepatterns
The file scan when member files are missing was broken due to earlier
changes to how pattern matching worked. The search should now correctly
identify the missing files. This also allowed the paring down to just a
single filepattern for GEFS input files.

Also, the regex in run_ncep_convert was changed to use the same pattern
rather than a hard-coded pattern.