Update linker to use GEFS_COM
The rocoto workflow already has to set the location of the GEFS COM
directory for the file dependencies of the link_input job, so instead
of using compath.py in the link_input script itself, the script is
updated to use GEFS_COM from the rocoto workflow. This means the
dependency and script are always consistent with one another and makes
it easy to change the location should a case need to be rerun from tape.