Add abililty to run using archived data
New scripts to retrieve the necessary data from the HPSS archive
in order to rerun a case that has already aged off the operational
stream.

A new task is added to run the new job, and an alternate workflow
(with associated cron script) add the new task to the workflow.
Importantly, the GEFS_COM in this alternate workflow is changed
from the operational location to PTMP.