Update link job for GEFS v12
The link job is updated to handle the new file pattern of GEFS v12.
Starting with GEFS v12, the 0.5-deg files will be used instead of
the discontinued 1-deg files.

In order to maintain backwards compatibility for reprocessing older
dates, the paths, filenames, and number of members are determined
based on the initialization time. Currently the transition time is
earlier than the operational switchover for testing purposes. This
will be changed to the actual operational cycle in a later commit.

Refs: #1