Update process tigge job for GEFS v12
The process tigge job is updated to handle the new filepatterns and
ensemble size of GEFS v12. Like previous commits, some constants had
to be replaced with functions to determine settings based on the
initialization time in order to maintain backwards compatability.

The TIGGE source code had to be substantially altered to handle
different ensemble sizes and resolution depending on the date. New
functions to calculate n_lat, n_lon, and n_points are added to handle
the resolution differences, as well as a get_n_members for ensemble
size and get_gen_proc_id for the generation process. Since the
resolution settings had previously been hard-coded into many of the
individual functions, many method signatures had to be updated to be
able to pass in additional needed resolution information.

Additionally, unused TIGGE source files are removed. Several files had
two versions, one with a 2016 added in the filename, with the 2016
version used. In a future commit, the filenames will be updated to
remove the year modifier. Code for test variables is removed.

The grib filter rules were altered to add one to the current ensemble
count instead of setting to a fixed number. This will correctly
update the count (TIGGE includes the control in the count) for any
ensemble size.

The output-verify list is replaced with a new series of nested for loops
that iterates through a multilevel hash. This approach is far more
maintainable than using a 75k+ line text file (whcih would've grown to
>100k for GEFS v12).

Cut-off dates for GEFS v12 settings is updated to the correct production date.

Due to the increased membership and resolution, the runtime had to be increased.

Refs: #2