Cleanup workflows
The PSLOT names (used in rocotoviewer) are updated for both workflows
to represent their data source.

Additionally, the hpss script was updated to have multiple cycledefs
(as an example) and the start/end variables used on prod (where cycles
are continuous) are removed. This will inform anyone new that they can
define multiple disconnected groups of cycles when running in historical
mode.