Delete lockfile when process task is rewound
If the process task is killed by the job scheduler (usually due to
LSF time limit or downtime), the lockfile would persist. Then when
the task is rewound to retry, the existence of the lockfile would
kill the rewound task and block the completion of the job. Now when
the task is rewound it will remove the lockfile so the task can
proceed without needing to delete the lockfile manually.