Improve handling of failures
A couple of changes to improve handling of failures. First, the
transfer job is separated into one for the rsync to rzdm and one
for archival to hpss. This allows one or the other to be retried
without impacting the other.

Also added a new job that will send an email if one of the jobs is
dead. This required the addition of a second new job to finalize
the cycle since the email job will only run if there is a failure.

Closes #6, #7